requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.3.2
pyyaml==6.0
python-dateutil==2.8.2
pytest==7.4.3
tqdm==4.66.1
